+++ primeSearchingUtilities.h
@@ -48,7 +48,7 @@ auto accumulate(BinaryOp op, S init, std::vector<T> factors) {
 template<typename T, typename G>
 auto applyPairwise(G g, std::vector<T> elements) {
     std::vector<decltype(g(elements[0], elements[0]))> results;
-    for (int i = 0; i < (elements.size() - 1); i++) {
+    for (size_t i = 0; i < (elements.size() - 1); i++) {
         results.push_back(g(elements[i], elements[i + 1]));
     }
     return results;
